[patch] refactor error dialog to index
