+++ web/src/components/modals/ErrorDialog/index.styles.ts
